fix(ui): make each terminal pane its own file drop target
Previously the OS file-drop listener was attached once to the entire
terminal content area, so a split view acted as a single drop zone and
the dropped path was always inserted into the globally active tab.

Scope dropping to each leaf panel instead: every pane gets its own
useOsFileDrop hook bound to its terminal area, and the dropped path is
sent to the session of the tab shown in that pane (resolved via the new
getPanelActiveSessionId helper). The drag highlight is now confined to
the pane under the cursor. Extracts quotePath into a shared util.
