+++ src/components/SplitView/SplitView.css
@@ -36,6 +36,24 @@
   overflow: hidden;
 }
 
+/* Per-pane OS file-drop highlight — scoped to the hovered terminal pane. */
+.split-view__file-drag-overlay {
+  position: absolute;
+  inset: 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: var(--spacing-sm);
+  background: rgba(0, 0, 0, 0.55);
+  border: 2px dashed var(--text-accent);
+  color: var(--text-accent);
+  font-size: var(--font-size-sm);
+  font-weight: 500;
+  pointer-events: none;
+  z-index: 100;
+}
+
 .split-view__empty {
   display: flex;
   align-items: center;
