+++ src/components/Terminal/TerminalView.css
@@ -52,20 +52,3 @@
   overflow: hidden;
   position: relative;
 }
-
-.terminal-view__drag-overlay {
-  position: absolute;
-  inset: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: var(--spacing-sm);
-  background: rgba(0, 0, 0, 0.55);
-  border: 2px dashed var(--text-accent);
-  color: var(--text-accent);
-  font-size: var(--font-size-sm);
-  font-weight: 500;
-  pointer-events: none;
-  z-index: 100;
-}
